Add AbstractFictitiousLfBranch
