+++ src/main/java/com/powsybl/openloadflow/network/impl/LfDanglingLineBranch.java
@@ -8,24 +8,15 @@ package com.powsybl.openloadflow.network.impl;
 
 import com.powsybl.iidm.network.DanglingLine;
 import com.powsybl.openloadflow.network.*;
-import com.powsybl.openloadflow.util.Evaluable;
 
 import java.util.Objects;
-
-import static com.powsybl.openloadflow.util.EvaluableConstants.NAN;
 
 /**
  * @author Geoffroy Jamgotchian <geoffroy.jamgotchian at rte-france.com>
  */
-public class LfDanglingLineBranch extends AbstractLfBranch {
+public class LfDanglingLineBranch extends AbstractFictitiousLfBranch {
 
     private final DanglingLine danglingLine;
-
-    private Evaluable p = NAN;
-
-    private Evaluable q = NAN;
-
-    private Evaluable i = NAN;
 
     protected LfDanglingLineBranch(LfNetwork network, LfBus bus1, LfBus bus2, PiModel piModel, DanglingLine danglingLine) {
         super(network, bus1, bus2, piModel);
@@ -56,66 +47,6 @@ public class LfDanglingLineBranch extends AbstractLfBranch {
     @Override
     public boolean hasPhaseControlCapability() {
         return false;
-    }
-
-    @Override
-    public void setP1(Evaluable p1) {
-        this.p = Objects.requireNonNull(p1);
-    }
-
-    @Override
-    public Evaluable getP1() {
-        return p;
-    }
-
-    @Override
-    public void setP2(Evaluable p2) {
-        // nothing to do
-    }
-
-    @Override
-    public Evaluable getP2() {
-        return NAN;
-    }
-
-    @Override
-    public void setQ1(Evaluable q1) {
-        this.q = Objects.requireNonNull(q1);
-    }
-
-    @Override
-    public Evaluable getQ1() {
-        return q;
-    }
-
-    @Override
-    public void setQ2(Evaluable q2) {
-        // nothing to do
-    }
-
-    @Override
-    public Evaluable getQ2() {
-        return NAN;
-    }
-
-    @Override
-    public void setI1(Evaluable i1) {
-        this.i = Objects.requireNonNull(i1);
-    }
-
-    @Override
-    public Evaluable getI1() {
-        return i;
-    }
-
-    @Override
-    public void setI2(Evaluable i2) {
-        // nothing to do
-    }
-
-    @Override
-    public Evaluable getI2() {
-        return NAN;
     }
 
     @Override
